fix: truncate the dashboard sparkline line to the content column
Every sibling line goes through ansi.Truncate, and this one did not. The
bar width stops shrinking at a floor of 8, so it is not always narrower
than the content column: with a 20 column pane and a 14 character pinned
label the line ran 26 columns wide and shifted the layout.

The spark-mode test asserted on the full block glyph, which renderBar
also draws, so it passed with no sparkline present. It now asserts on the
series minimum, which only the sparkline alphabet produces. The three
sparkline tests move to their own file, since the dashboard test file
reached the 800 line cap.
