+++ internal/app/commands_dashboard_test.go
@@ -778,45 +778,3 @@ func TestComposeDashboard_ClusterSeriesKeyedByPreviewTarget(t *testing.T) {
 	assert.NotEqual(t, stripANSI(contentA), stripANSI(contentB),
 		"the cluster sparkline must reflect the previewed member's own history")
 }
-
-func TestDashboardResourcesSection_DrawsSparklineInSparkMode(t *testing.T) {
-	data := dashboardData{
-		totalCPUUsed: 1000, totalCPUAlloc: 4000,
-		totalMemUsed: 1024 * 1024 * 512, totalMemAlloc: 1024 * 1024 * 1024,
-		cpuSeries: k8s.MetricSeries{Points: []float64{1, 5, 9}},
-		memSeries: k8s.MetricSeries{Points: []float64{9, 5, 1}},
-	}
-	// content must be set: dashboardMetricLines truncates every line to it,
-	// so a zero value renders empty lines regardless of the sparkline.
-	w := dashboardWidths{bar: 20, node: 20, sep: 60, label: 6, content: 80}
-
-	lines := dashboardResourcesSection(nil, data, w, true)
-
-	joined := stripANSI(strings.Join(lines, "\n"))
-	assert.Contains(t, joined, "█", "a series reaching its maximum must draw the top glyph")
-	assert.Contains(t, joined, "1.0 / 4.0", "the numeric total must still be shown")
-}
-
-func TestDashboardResourcesSection_NumericModeDrawsNoGlyphs(t *testing.T) {
-	data := dashboardData{
-		totalCPUUsed: 1000, totalCPUAlloc: 4000,
-		cpuSeries: k8s.MetricSeries{Points: []float64{1, 5, 9}},
-	}
-	// content must be set: dashboardMetricLines truncates every line to it,
-	// so a zero value renders empty lines regardless of the sparkline.
-	w := dashboardWidths{bar: 20, node: 20, sep: 60, label: 6, content: 80}
-
-	lines := dashboardResourcesSection(nil, data, w, false)
-
-	joined := stripANSI(strings.Join(lines, "\n"))
-	// "█" is excluded: renderBar's fill uses it too, and a nonzero usage bar
-	// (a more realistic fixture than an all-empty one) draws it regardless of
-	// sparklines. The other seven glyphs are unique to the sparkline alphabet.
-	for _, g := range ui.SparklineGlyphs {
-		if g == '█' {
-			continue
-		}
-		assert.NotContains(t, joined, string(g),
-			"numeric mode must render identically to before this feature")
-	}
-}
